Move NemLogIn button outside of module
